+++ lib/EasyPost/Webhook.php
@@ -9,7 +9,6 @@ namespace EasyPost;
  * @property string $modoe
  * @property string $url
  * @property string $disabled_at
+ * @property array $custom_headers
  */
-class Webhook extends EasyPostObject
-{
-}
+class Webhook extends EasyPostObject {}