+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## Next Release
 
+- Adds `custom_headers` property to webhook model
 - Corrects wrapping payload for update webhook endpoint
 - Removes the deprecated `create_list` tracker endpoint function as it is no longer available via API